# backend/docs/swagger.yaml
basePath: /
definitions:
...
  dehydrotationlab4_internal_app_ds.User:
    properties:
      id:
        type: integer
      is_moderator:
        type: boolean
      login:
        type: string
    type: object
...
host: localhost:8082
info:
  contact:
    email: support@partitionlab.com
    name: API Support
  description: |-
    API для управления заявками на расчет звукоизоляции перегородок

    Без авторизации доступны:
    - Регистрация и аутентификация
    - Просмотр симптомов (GET)

    С авторизацией (пользователь):
    - Управление своими заявками
    - Добавление симптомов в заявки

    С ролью модератора:
    - Все методы пользователя
    - CRUD операции с симптомами
    - Завершение/отклонение заявок
  title: Partition Lab API
  version: "4.0"
paths:
...
  /api/users/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные для входа
        in: body
        name: request
        required: true
        schema:
          properties:
            login:
              type: string
            password:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              session_id:
                type: string
              token:
                type: string
              user:
                $ref: '#/definitions/dehydrotationlab4_internal_app_ds.User'
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Вход пользователя
      tags:
      - auth
  /api/users/logout:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
            type: object
      security:
      - BearerAuth: []
      - CookieAuth: []
      summary: Выход пользователя
      tags:
      - auth
  /api/users/profile:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              user:
                $ref: '#/definitions/dehydrotationlab4_internal_app_ds.User'
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      - CookieAuth: []
      summary: Получение профиля текущего пользователя
      tags:
      - auth
    put:
      consumes:
      - application/json
      parameters:
      - description: Новые данные профиля
        in: body
        name: request
        required: true
        schema:
          properties:
            login:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              user:
                $ref: '#/definitions/dehydrotationlab4_internal_app_ds.User'
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      - CookieAuth: []
      summary: Обновление профиля
      tags:
      - auth
  /api/users/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные для регистрации
        in: body
        name: request
        required: true
        schema:
          properties:
            login:
              type: string
            password:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              user:
                $ref: '#/definitions/dehydrotationlab4_internal_app_ds.User'
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Регистрация нового пользователя
      tags:
      - auth
securityDefinitions:
  BearerAuth:
    description: 'Введите JWT токен в формате: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
  CookieAuth:
    description: Сессия через cookie (session_id)
    in: header
    name: Cookie
    type: apiKey
swagger: "2.0"